refactor: split CLI hook handlers into per-agent modules
Extract shared hook logic (git info, env vars, DB insertion, result
output) into hooks/mod.rs, and move each agent's handler into its own
file (claude.rs, codex.rs, opencode.rs). Agent-specific logic such as
stdin reading, body truncation, and event-type mapping remain in the
respective files. Introduce NotificationPayload struct to group DB
insert parameters and satisfy clippy's too-many-arguments lint.
